==> 1. Data Cleaning.ipynb <==
# Political Social Media Analysis

In this project, I will try to compare the tweets of Donald Trump, Barrack Obama, and Hillary Clinton to come up with meaningful insights

Data:
There are 3 CSV files which will be used:
1. DonaldTrump
2. BarackObama
3. HillaryClinton

All 3 have the same structure
sl no,date,id,link,retweet,text,author

Import libraries

import pandas as pd
import re

Read the data

trump = pd.read_csv("data/DonaldTrump.csv")
obama = pd.read_csv("data/BarackObama.csv")
clinton = pd.read_csv("data/HillaryClinton.csv")

### Check out the data

print("Number of tweets by Trump: ", trump.shape[0])
print("Number of tweets by Obama: ", obama.shape[0])
print("Number of tweets by Hillary: ", clinton.shape[0])

trump.head()

obama.head()

clinton.head()

## Data Cleaning

A couple of issues have been identified by eye-balling.
1. There is an unnamed column which is unnecessary and must be dropped
2. ID, and link are not required as well
3. Date format is not standard
4. Since this is a tweet, it contains many non-alphanumeric characters as well

#Dropping the unnecessary columns
drop_cols = ["Unnamed: 0", "id", "link"]
trump.drop(drop_cols, axis=1, inplace=True)
obama.drop(drop_cols, axis=1, inplace=True)
clinton.drop(drop_cols, axis=1, inplace=True)

For our analysis, we do not need non-alphanumeric characters like "%^&*() etc"  
We can also remove all references to images like "pic.twitter.com."

def replaceUnnecessaryText(text):
    text = text.split('pic.twitter')[0]   #We can also remove the references to an image
    text_list = text.split()
    text=""
    for word in text_list:
        word = re.sub(r'@\S+', '', word)     #Remove all words beginning with @ to remove references to people
        text+=" "+word
    return re.sub('[^a-zA-Z0-9#@.,/?! ]', '', text)


trump['text'] = trump['text'].apply(replaceUnnecessaryText)
obama['text'] = obama['text'].apply(replaceUnnecessaryText)
clinton['text'] = clinton['text'].apply(replaceUnnecessaryText)

Now we will work on standardising the date format

Some dates have years while some don't.
The ones who don't are of the year 2016. Hence, we will first attach 2017 to the date

def changeDateFormat(date):
    new_date = date
    if(len(date)<7):                        #Date doesn't have year attached yet
        if('-' in date):
            date_arr = date.split('-')         #Date can be separated by '-' or ' '
        else:            
            date_arr = date.split()            #Add year 2016 to the dates with no year.
        new_date = date_arr[1]+" "+date_arr[0]+" 2016"
    return new_date

trump.date = trump.date.apply(changeDateFormat)
obama.date = obama.date.apply(changeDateFormat)
clinton.date = clinton.date.apply(changeDateFormat)

Convert the string to a datetime format

trump['date'] = pd.to_datetime(trump['date'])
obama['date'] = pd.to_datetime(obama['date'])
clinton['date'] = pd.to_datetime(clinton['date'])

All Trump and Clinton tweets are from 01-01-2014 while Obama tweets start way earlier in 20017.

Remove all pre-2014 Obama tweets

obama = obama[obama['date'].dt.year > 2013]

#By printing the tail, we can confirm that all pre-2014 tweets have been dropped
obama.tail()

Now let's compare the length of all 3 again

print("Number of tweets by Trump: ", trump.shape[0])
print("Number of tweets by Obama: ", obama.shape[0])
print("Number of tweets by Hillary: ", clinton.shape[0])

trump.head()

## Writing to CSV files

Use index=False so that you don't save the first unnamed column in the CSV file

trump.to_csv("data/DonaldTrumpClean.csv", index=False)
obama.to_csv("data/BarackObamaClean.csv", index=False)
clinton.to_csv("data/HillaryClintonClean.csv", index=False)

## References

Data collected from: https://www.kaggle.com/nandys/social-media-analysis-kim-kardashian/data
